+++ e2e/evm/test/debug_queries.test.ts
@@ -32,6 +32,11 @@ describe("debug queries", () => {
   it("debug_traceBlockByNumber", async () => {
     const traceResult = await provider.send("debug_traceBlockByNumber", [
       blockNumber,
+      {
+        tracer: "callTracer",
+        timeout: "3000s",
+        tracerConfig: { onlyTopCall: false },
+      },
     ])
     expectTrace(traceResult)
   })
@@ -39,13 +44,24 @@ describe("debug queries", () => {
   it("debug_traceBlockByHash", async () => {
     const traceResult = await provider.send("debug_traceBlockByHash", [
       blockHash,
+      {
+        tracer: "callTracer",
+        timeout: "3000s",
+        tracerConfig: { onlyTopCall: false },
+      },
     ])
     expectTrace(traceResult)
   })
 
   it("debug_traceTransaction", async () => {
-    // it.skip("debug_traceTransaction", async () => {
-    const traceResult = await provider.send("debug_traceTransaction", [txHash])
+    const traceResult = await provider.send("debug_traceTransaction", [
+      txHash,
+      {
+        tracer: "callTracer",
+        timeout: "3000s",
+        tracerConfig: { onlyTopCall: false },
+      },
+    ])
     expectTrace([{ result: traceResult }])
   })
 
@@ -60,7 +76,11 @@ describe("debug queries", () => {
     const traceResult = await provider.send("debug_traceCall", [
       tx,
       "latest",
-      {},
+      {
+        tracer: "callTracer",
+        timeout: "3000s",
+        tracerConfig: { onlyTopCall: false },
+      },
     ])
     expectTrace([{ result: traceResult }])
   })
@@ -103,8 +123,11 @@ const expectTrace = (traceResult: any[]) => {
   expect(traceResult.length).toBeGreaterThan(0)
 
   const trace = traceResult[0]["result"]
-  expect(trace).toHaveProperty("failed", false)
+  expect(trace).toHaveProperty("from")
+  expect(trace).toHaveProperty("to")
   expect(trace).toHaveProperty("gas")
-  expect(trace).toHaveProperty("returnValue")
-  expect(trace).toHaveProperty("structLogs")
+  expect(trace).toHaveProperty("gasUsed")
+  expect(trace).toHaveProperty("input")
+  expect(trace).toHaveProperty("output")
+  expect(trace).toHaveProperty("type", "CALL")
 }